Improved additional files and fixed some issues

+++ src/c3/create_operator.py
@@ -5,6 +5,8 @@ import logging
 import shutil
 import argparse
 import subprocess
+import glob
+import re
 from pathlib import Path
 from string import Template
 from typing import Optional
@@ -18,14 +20,31 @@ from c3.templates import (python_component_setup_code, r_component_setup_code,
 CLAIMED_VERSION = 'V0.1'
 
 
-def create_dockerfile(dockerfile_template, requirements, target_code, additional_files_path, working_dir, command):
+def create_dockerfile(dockerfile_template, requirements, target_code, target_dir, additional_files, working_dir, command):
+    # Check for requirements file
+    for i in range(len(requirements)):
+        if '-r ' in requirements[i]:
+            r_file_search = re.search('-r ~?\/?([A-Za-z0-9\/]*\.txt)', requirements[i])
+            if len(r_file_search.groups()):
+                # Get file from regex
+                requirements_file = r_file_search.groups()[0]
+                if requirements_file not in additional_files and os.path.isfile(requirements_file):
+                    # Add missing requirements text file to additional files
+                    additional_files.append(r_file_search.groups()[0])
+            if '/' not in requirements[i]:
+                # Add missing home directory to the command `pip install -r ~/requirements.txt`
+                requirements[i] = requirements[i].replace('-r ', '-r ~/')
+
     requirements_docker = list(map(lambda s: 'RUN ' + s, requirements))
     requirements_docker = '\n'.join(requirements_docker)
+    additional_files_docker = list(map(lambda s: f"ADD {s} {working_dir}{s}", additional_files))
+    additional_files_docker = '\n'.join(additional_files_docker)
 
     docker_file = dockerfile_template.substitute(
         requirements_docker=requirements_docker,
         target_code=target_code,
-        additional_files_path=additional_files_path,
+        target_dir=target_dir,
+        additional_files_docker=additional_files_docker,
         working_dir=working_dir,
         command=os.path.basename(command),
     )
@@ -33,9 +52,10 @@ def create_dockerfile(dockerfile_template, requirements, target_code, additional
     logging.info('Create Dockerfile')
     with open("Dockerfile", "w") as text_file:
         text_file.write(docker_file)
+    logging.debug('Dockerfile:\n' + docker_file)
 
 
-def create_kfp_component(name, description, repository, version, command, target_code, file_path, inputs, outputs):
+def create_kfp_component(name, description, repository, version, command, target_code, target_dir, file_path, inputs, outputs):
 
     inputs_list = str()
     for name, options in inputs.items():
@@ -67,7 +87,10 @@ def create_kfp_component(name, description, repository, version, command, target
         version=version,
         inputs=inputs_list,
         outputs=outputs_list,
-        call=f'{os.path.basename(command)} ./{target_code} {parameter_list}',
+        command=os.path.basename(command),
+        target_dir=target_dir,
+        target_code=target_code,
+        parameter_list=parameter_list,
         parameter_values=parameter_values,
     )
 
@@ -79,7 +102,7 @@ def create_kfp_component(name, description, repository, version, command, target
         text_file.write(yaml)
 
 
-def create_kubernetes_job(name, repository, version, target_code, command, working_dir, file_path, inputs):
+def create_kubernetes_job(name, repository, version, target_code, target_dir, command, working_dir, file_path, inputs):
     # get environment entries
     env_entries = str()
     for key in list(inputs.keys()):
@@ -91,6 +114,7 @@ def create_kubernetes_job(name, repository, version, target_code, command, worki
         repository=repository,
         version=version,
         target_code=target_code,
+        target_dir=target_dir,
         env_entries=env_entries,
         command=command,
         working_dir=working_dir,
@@ -138,6 +162,46 @@ def create_cwl_component(name, repository, version, file_path, inputs, outputs):
         text_file.write(cwl)
 
 
+def check_existing_files(file_path, rename_files, overwrite_files):
+    if rename_files is None and overwrite_files:
+        # Overwrite potential files
+        return
+
+    target_job_yaml_path = Path(file_path).with_suffix('.job.yaml')
+
+    # Check for existing job yaml
+    if target_job_yaml_path.is_file():
+        if rename_files is None:
+            # Ask user
+            rename_files = input(f'\nFound a existing Kubernetes job file at {target_job_yaml_path}.\n'                      
+                                 f'ENTER to overwrite the file, write Y to rename the file to '
+                                 f'modified_{target_job_yaml_path.name}, or provide a custom name:\n')
+        if rename_files.strip() == '':
+            # Overwrite file
+            return
+        elif rename_files.lower() == 'y':
+            # Default file name
+            new_file_name = 'modified_' + Path(file_path).name
+        else:
+            # Rename to custom name
+            new_file_name = rename_files
+
+        modified_path = (target_job_yaml_path.parent / new_file_name).with_suffix('.job.yaml')
+        # Check if modified path exists and potentially overwrite
+        if modified_path.exists():
+            if overwrite_files:
+                logging.info(f'Overwriting modified path {modified_path}.')
+            else:
+                overwrite = input(f'Modified path {modified_path} already exists. ENTER to overwrite the file.')
+                if overwrite != '':
+                    logging.error(f'Abort creating operator. Please rename file manually and rerun the script.')
+                    raise FileExistsError
+
+        os.rename(str(target_job_yaml_path), str(modified_path))
+        logging.info(f'Renamed Kubernetes job file to {modified_path}')
+    # TODO: Should we check other files too? Currently assuming no modification for yaml and cwl.
+
+
 def print_claimed_command(name, repository, version, inputs):
     claimed_command = f"claimed --component {repository}/claimed-{name}:{version}"
     for input, options in inputs.items():
@@ -145,13 +209,13 @@ def print_claimed_command(name, repository, version, inputs):
     logging.info(f'Run operators locally with claimed-cli:\n{claimed_command}')
 
 
-def remove_temporary_files(file_path, target_code, additional_files_path):
+def remove_temporary_files(file_path, target_code):
     logging.info(f'Remove local files')
     # remove temporary files
     if file_path != target_code:
         os.remove(target_code)
-    os.remove('Dockerfile')
-    shutil.rmtree(additional_files_path, ignore_errors=True)
+    if os.path.isfile('Dockerfile'):
+        os.remove('Dockerfile')
 
 
 def create_operator(file_path: str,
@@ -162,6 +226,8 @@ def create_operator(file_path: str,
                     log_level='INFO',
                     test_mode=False,
                     no_cache=False,
+                    rename_files=None,
+                    overwrite_files=False,
                     ):
     logging.info('Parameters: ')
     logging.info('file_path: ' + file_path)
@@ -176,20 +242,16 @@ def create_operator(file_path: str,
         working_dir = '/opt/app-root/src/'
 
     elif file_path.endswith('.py'):
-        target_code = file_path.split('/')[-1]
-        if file_path == target_code:
-            # use temp file for processing
-            target_code = 'claimed_' + target_code
+        # use temp file for processing
+        target_code = 'claimed_' + os.path.basename(file_path)
         # Copy file to current working directory
         shutil.copy(file_path, target_code)
         command = '/opt/app-root/bin/python'
         working_dir = '/opt/app-root/src/'
 
     elif file_path.lower().endswith('.r'):
-        target_code = file_path.split('/')[-1]
-        if file_path == target_code:
-            # use temp file for processing
-            target_code = 'claimed_' + target_code
+        # use temp file for processing
+        target_code = 'claimed_' + os.path.basename(file_path)
         # Copy file to current working directory
         shutil.copy(file_path, target_code)
         command = 'Rscript'
@@ -232,27 +294,41 @@ def create_operator(file_path: str,
     # Strip 'claimed-' from name of copied temp file
     if name.startswith('claimed-'):
         name = name[8:]
+    target_dir = os.path.dirname(file_path)
+    # Check that the main file is within the cwd
+    if '../' in target_dir:
+        raise PermissionError(f"Forbidden path outside the docker build context: {target_dir}. "
+                              f"Change the current working directory to include the file.")
+    elif target_dir != '':
+        target_dir += '/'
 
     logging.info('Operator name: ' + name)
     logging.info('Description:: ' + description)
     logging.info('Inputs: ' + str(inputs))
     logging.info('Outputs: ' + str(outputs))
     logging.info('Requirements: ' + str(requirements))
+    logging.debug(f'Target code: {target_code}')
+    logging.debug(f'Target directory: {target_dir}')
 
-    # copy all additional files to temporary folder
-    additional_files_path = 'additional_files_path'
-    while os.path.exists(additional_files_path):
-        # ensures using a new directory
-        additional_files_path += '_temp'
-    logging.debug(f'Create dir for additional files {additional_files_path}')
-    os.makedirs(additional_files_path)
-    for additional_file in additional_files:
-        assert os.path.isfile(additional_file), \
-            f"Could not find file at {additional_file}. Please provide only files as additional parameters."
-        shutil.copy(additional_file, additional_files_path)
-    logging.info(f'Selected additional files: {os.listdir(additional_files_path)}')
+    # Load all additional files
+    logging.debug('Looking for additional files:')
+    additional_files_found = []
+    for file_pattern in additional_files:
+        if '../' in file_pattern:
+            # Check that additional file are within the cwd
+            raise PermissionError(f"Forbidden path outside the docker build context: {file_pattern}. "
+                                  f"Change the current working directory to include all additional files.")
+        # Include files based on wildcards
+        files_found = glob.glob(file_pattern)
+        if len(files_found) == 0:
+            raise FileNotFoundError(f'No additional files for path {file_pattern}.')
+        additional_files_found.extend(files_found)
+        logging.debug(f'Searched for "{file_pattern}". Found {", ".join(files_found)}')
+    logging.info(f'Found {len(additional_files_found)} additional files and directories\n'
+                 f'{", ".join(additional_files_found)}')
 
-    create_dockerfile(dockerfile_template, requirements, target_code, additional_files_path, working_dir, command)
+    create_dockerfile(dockerfile_template, requirements, target_code, target_dir, additional_files_found, working_dir,
+                      command)
 
     if version is None:
         # auto increase version based on registered images
@@ -277,7 +353,8 @@ def create_operator(file_path: str,
             stdout=None if log_level == 'DEBUG' else subprocess.PIPE, check=True, shell=True,
         )
     except Exception as err:
-        remove_temporary_files(file_path, target_code, additional_files_path)
+        remove_temporary_files(file_path, target_code)
+        logging.error('Docker build failed. Consider running C3 with `--log_level DEBUG` to see the docker build logs.')
         raise err
     logging.info('Successfully built image')
 
@@ -300,20 +377,28 @@ def create_operator(file_path: str,
             logging.info('Continue processing (test mode).')
             pass
         else:
-            remove_temporary_files(file_path, target_code, additional_files_path)
+            remove_temporary_files(file_path, target_code)
             raise err
 
-    # Create application scripts
-    create_kfp_component(name, description, repository, version, command, target_code, file_path, inputs, outputs)
+    # Check for existing files and optionally modify them before overwriting
+    try:
+        check_existing_files(file_path, rename_files, overwrite_files)
+    except Exception as err:
+        remove_temporary_files(file_path, target_code)
+        raise err
 
-    create_kubernetes_job(name, repository, version, target_code, command, working_dir, file_path, inputs)
+    # Create application scripts
+    create_kfp_component(name, description, repository, version, command, target_code, target_dir, file_path, inputs,
+                         outputs)
+
+    create_kubernetes_job(name, repository, version, target_code, target_dir, command, working_dir, file_path, inputs)
 
     create_cwl_component(name, repository, version, file_path, inputs, outputs)
 
     print_claimed_command(name, repository, version, inputs)
 
     # Remove temp files
-    remove_temporary_files(file_path, target_code, additional_files_path)
+    remove_temporary_files(file_path, target_code)
 
 
 def main():
@@ -326,6 +411,9 @@ def main():
                         help='Container registry address, e.g. docker.io/<username>')
     parser.add_argument('-v', '--version', type=str, default=None,
                         help='Container image version. Auto-increases the version number if not provided (default 0.1)')
+    parser.add_argument('--rename', type=str, nargs='?', default=None, const='',
+                        help='Rename existing yaml files (argument without value leads to modified_{file name})')
+    parser.add_argument('--overwrite', action='store_true', help='Overwrite existing yaml files')
     parser.add_argument('-l', '--log_level', type=str, default='INFO')
     parser.add_argument('--dockerfile_template_path', type=str, default='',
                         help='Path to custom dockerfile template')
@@ -359,6 +447,8 @@ def main():
         log_level=args.log_level,
         test_mode=args.test_mode,
         no_cache=args.no_cache,
+        overwrite_files=args.overwrite,
+        rename_files=args.rename,
     )
